+++ src/ParametersHandler/tests/ParametersHandlerTest.cpp
@@ -14,61 +14,224 @@
 // Catch2
 #include <catch2/catch.hpp>
 
+#include <BipedalLocomotionControllers/GenericContainer/TemplateHelpers.h>
+#include <BipedalLocomotionControllers/GenericContainer/Vector.h>
 #include <BipedalLocomotionControllers/ParametersHandler/IParametersHandler.h>
 
 using namespace BipedalLocomotionControllers::ParametersHandler;
+using namespace BipedalLocomotionControllers;
 
 // example of parameters handler
-class BasicImplementation : public IParametersHandler<BasicImplementation>
+class BasicImplementation : public IParametersHandler
 {
     std::unordered_map<std::string, std::any> m_map;
 
-public:
-    BasicImplementation(const std::unordered_map<std::string, std::any>& map)
-        : m_map{map}
-    {}
-
-    BasicImplementation() = default;
-
-    template <typename T> void setParameter(const std::string& parameterName, const T& parameter)
+    template <typename T>
+    void setParameterPrivate(const std::string& parameterName, const T& parameter)
     {
-        m_map[parameterName] = std::make_any<T>(parameter);
+        // a scalar element and a strings is retrieved using getElementFromSearchable() function
+        if constexpr (std::is_scalar<T>::value || is_string<T>::value)
+            m_map[parameterName] = parameter;
+        else
+        {
+            using elementType = typename T::value_type;
+            std::vector<elementType> tempParameter(parameter.size());
+
+            for (std::size_t index = 0; index < parameter.size(); index++)
+                tempParameter[index] = parameter[index];
+
+            m_map[parameterName] = std::make_any<std::vector<elementType>>(tempParameter);
+        }
     }
 
-    template <typename T> bool getParameter(const std::string& parameterName, T& parameter) const
+    template <typename T>
+    bool getParameterPrivate(const std::string& parameterName, T& parameter) const
     {
         auto parameterAny = m_map.find(parameterName);
         if (parameterAny == m_map.end())
         {
-            std::cerr << "[BasicImplementation::getParameter] Parameter named " << parameterName
-                      << " not found." << std::endl;
+            std::cerr << "[BasicImplementation::getParameterPrivate] Parameter named "
+                      << parameterName << " not found." << std::endl;
             return false;
         }
 
-        try
+        if constexpr (std::is_scalar<T>::value || is_string<T>::value)
         {
-            parameter = std::any_cast<T>(parameterAny->second);
-        } catch (const std::bad_any_cast& exception)
+            try
+            {
+                parameter = std::any_cast<T>(parameterAny->second);
+            } catch (const std::bad_any_cast& exception)
+            {
+                std::cerr << "[BasicImplementation::getParameterPrivate] The type of the parameter "
+                             "named "
+                          << parameterName << " is different from the one expected" << std::endl;
+                return false;
+            }
+        } else
         {
-            std::cerr << "[BasicImplementation::getParameter] The type of the parameter named "
-                      << parameterName << " is different from the one expected" << std::endl;
+            using elementType = typename T::value_type;
+            std::vector<elementType> castedParameter;
+            try
+            {
+                castedParameter = std::any_cast<std::vector<elementType>>(parameterAny->second);
+            } catch (const std::bad_any_cast& exception)
+            {
+                std::cerr << "[BasicImplementation::getParameterPrivate] The type of the parameter "
+                             "named "
+                          << parameterName << " is different from the one expected" << std::endl;
+                return false;
+            }
+
+            if (castedParameter.size() != parameter.size())
+            {
+                // If the vector can be resize, let resize it. Otherwise it is a fix-size vector and
+                // the dimensions has to be the same of list
+                if constexpr (GenericContainer::is_vector<T>::value)
+                {
+                    if (!parameter.resizeVector(castedParameter.size()))
+                    {
+                        std::cerr << "[BasicImplementation::getParameterPrivate] Unable to resize "
+                                  << type_name<T>() << "List size: " << castedParameter.size()
+                                  << ". Vector size: " << parameter.size() << std::endl;
+                        return false;
+                    }
+                } else if constexpr (is_resizable<T>::value)
+                    parameter.resize(castedParameter.size());
+                else
+                {
+                    std::cerr << "[BasicImplementation::getParameterPrivate] The size of the "
+                                 "vector does not match with the size of the list. List size: "
+                              << castedParameter.size() << ". Vector size: " << parameter.size()
+                              << std::endl;
+                    return false;
+                }
+            }
+
+            for (std::size_t index = 0; index < parameter.size(); index++)
+                parameter[index] = castedParameter[index];
+        }
+        return true;
+    }
+
+    void setParameter(const std::string& parameterName,
+                      const GenericContainer::Vector<const int>& parameter) final
+    {
+        return setParameterPrivate(parameterName, parameter);
+    }
+    void setParameter(const std::string& parameterName,
+                      const GenericContainer::Vector<const double>& parameter) final
+    {
+        return setParameterPrivate(parameterName, parameter);
+    }
+
+    void setParameter(const std::string& parameterName,
+                      const GenericContainer::Vector<const std::string>& parameter) final
+    {
+        return setParameterPrivate(parameterName, parameter);
+    }
+
+    bool getParameter(const std::string& parameterName,
+                      GenericContainer::Vector<int>& parameter) const final
+    {
+        return getParameterPrivate(parameterName, parameter);
+    }
+
+    bool getParameter(const std::string& parameterName,
+                      GenericContainer::Vector<double>& parameter) const final
+    {
+        return getParameterPrivate(parameterName, parameter);
+    }
+
+    bool getParameter(const std::string& parameterName,
+                      GenericContainer::Vector<std::string>& parameter) const final
+    {
+        return getParameterPrivate(parameterName, parameter);
+    }
+
+public:
+    BasicImplementation(const std::unordered_map<std::string, std::any>& map)
+        : m_map{map}
+    {
+    }
+
+    BasicImplementation() = default;
+
+    bool getParameter(const std::string& parameterName, int& parameter) const final
+    {
+        return getParameterPrivate(parameterName, parameter);
+    }
+
+    bool getParameter(const std::string& parameterName, double& parameter) const final
+    {
+        return getParameterPrivate(parameterName, parameter);
+    }
+
+    bool getParameter(const std::string& parameterName, std::string& parameter) const final
+    {
+        return getParameterPrivate(parameterName, parameter);
+    }
+
+    bool getParameter(const std::string& parameterName, bool& parameter) const final
+    {
+        return getParameterPrivate(parameterName, parameter);
+    }
+
+    bool getParameter(const std::string& parameterName, std::vector<bool>& parameter) const final
+    {
+        return getParameterPrivate(parameterName, parameter);
+    }
+
+    void setParameter(const std::string& parameterName, const int& parameter) final
+    {
+        return setParameterPrivate(parameterName, parameter);
+    }
+
+    void setParameter(const std::string& parameterName, const double& parameter) final
+    {
+        return setParameterPrivate(parameterName, parameter);
+    }
+
+    void setParameter(const std::string& parameterName, const char* parameter) final
+    {
+        return setParameterPrivate(parameterName, std::string(parameter));
+    }
+
+    void setParameter(const std::string& parameterName, const std::string& parameter) final
+    {
+        return setParameterPrivate(parameterName, parameter);
+    }
+
+    void setParameter(const std::string& parameterName, const bool& parameter) final
+    {
+        return setParameterPrivate(parameterName, parameter);
+    }
+
+    void setParameter(const std::string& parameterName, const std::vector<bool>& parameter) final
+    {
+        return setParameterPrivate(parameterName, parameter);
+    }
+
+    bool setGroup(const std::string& name, IParametersHandler::shared_ptr newGroup)
+    {
+        auto downcastedPtr = std::dynamic_pointer_cast<BasicImplementation>(newGroup);
+        if (downcastedPtr == nullptr)
+        {
+            std::cerr << "[BasicImplementation::setGroup] Unable to downcast the pointer to "
+                         "BasicImplementation."
+                      << std::endl;
             return false;
         }
+
+        m_map[name] = std::make_any<shared_ptr>(downcastedPtr);
 
         return true;
     }
 
-    void setGroup(const std::string& name, shared_ptr newGroup)
-    {
-        m_map[name] = std::make_any<shared_ptr>(newGroup);
-    }
-
-
-    BasicImplementation::weak_ptr getGroup(const std::string& name) const
+    IParametersHandler::weak_ptr getGroup(const std::string& name) const final
     {
         auto group = m_map.find(name);
         if (group == m_map.end())
-            return BasicImplementation::make_shared();
+            return std::make_shared<BasicImplementation>();
 
         shared_ptr map;
         try
@@ -78,7 +241,7 @@ public:
         {
             std::cerr << "[BasicImplementation::getGroup] The element named " << name
                       << " is not a 'std::unordered_map<std::string, std::any>'" << std::endl;
-            return BasicImplementation::make_shared();
+            return std::make_shared<BasicImplementation>();
         }
 
         return map;
@@ -89,21 +252,21 @@ public:
         m_map = object;
     }
 
-    std::string toString() const
+    std::string toString() const final
     {
         std::string key;
-        for (const auto& parameters: m_map)
+        for (const auto& parameters : m_map)
             key += parameters.first + " ";
 
         return key;
     }
 
-    bool isEmpty() const
+    bool isEmpty() const final
     {
         return m_map.size() == 0;
     }
 
-    void clear()
+    void clear() final
     {
         m_map.clear();
     }
@@ -113,11 +276,13 @@ public:
 
 TEST_CASE("Get parameters")
 {
-    BasicImplementation::unique_ptr parameterHandler = BasicImplementation::make_unique();
+    std::shared_ptr<BasicImplementation> originalHandler = std::make_shared<BasicImplementation>();
+    IParametersHandler::shared_ptr parameterHandler = originalHandler;
+
     parameterHandler->setParameter("answer_to_the_ultimate_question_of_life", 42);
     parameterHandler->setParameter("pi", 3.14);
     parameterHandler->setParameter("Fibonacci Numbers", std::vector<int>{1, 1, 2, 3, 5, 8, 13, 21});
-    parameterHandler->setParameter("John", std::string("Smith"));
+    parameterHandler->setParameter("John", "Smith");
 
     SECTION("Get integer")
     {
@@ -143,28 +308,34 @@ TEST_CASE("Get parameters")
     SECTION("Get Vector")
     {
         std::vector<int> element;
-        REQUIRE(parameterHandler->getParameter("Fibonacci Numbers", element));
+        REQUIRE(parameterHandler->getParameter("Fibonacci Numbers",
+                                               element,
+                                               GenericContainer::VectorResizeMode::Resizable));
         REQUIRE(element == std::vector<int>{1, 1, 2, 3, 5, 8, 13, 21});
     }
 
     SECTION("Set/Get Group")
     {
-        BasicImplementation::shared_ptr newGroup = BasicImplementation::make_shared();
-        parameterHandler->setGroup("CARTOONS", newGroup);
-        BasicImplementation::shared_ptr groupHandler = parameterHandler->getGroup("CARTOONS").lock();
+        BasicImplementation::shared_ptr newGroup = std::make_shared<BasicImplementation>();
+        REQUIRE(parameterHandler->setGroup("CARTOONS", newGroup));
+        BasicImplementation::shared_ptr groupHandler
+            = parameterHandler->getGroup("CARTOONS").lock();
         REQUIRE(groupHandler);
         groupHandler->setParameter("Donald's nephews",
                                    std::vector<std::string>{"Huey", "Dewey", "Louie"});
         std::vector<std::string> element;
-        REQUIRE(groupHandler->getParameter("Donald's nephews", element));
+        REQUIRE(groupHandler->getParameter("Donald's nephews",
+                                           element,
+                                           GenericContainer::VectorResizeMode::Resizable));
         REQUIRE(element == std::vector<std::string>{"Huey", "Dewey", "Louie"});
     }
 
     SECTION("is Empty")
     {
-        BasicImplementation::shared_ptr newGroup = BasicImplementation::make_shared();
-        parameterHandler->setGroup("CARTOONS", newGroup);
-        BasicImplementation::shared_ptr groupHandler = parameterHandler->getGroup("CARTOONS").lock();
+        BasicImplementation::shared_ptr newGroup = std::make_shared<BasicImplementation>();
+        REQUIRE(parameterHandler->setGroup("CARTOONS", newGroup));
+        BasicImplementation::shared_ptr groupHandler
+            = parameterHandler->getGroup("CARTOONS").lock();
         REQUIRE(groupHandler);
         REQUIRE(groupHandler->isEmpty());
 
@@ -176,14 +347,14 @@ TEST_CASE("Get parameters")
 
     SECTION("Print content")
     {
-        std::cout << "Parameters: " << *parameterHandler << std::endl;
+        std::cout << "Parameters: " << parameterHandler->toString() << std::endl;
     }
 
     SECTION("Set from object")
     {
         std::unordered_map<std::string, std::any> object;
         object["value"] = std::make_any<int>(10);
-        parameterHandler->set(object);
+        originalHandler->set(object);
         int expected;
         REQUIRE(parameterHandler->getParameter("value", expected));
         REQUIRE(expected == 10);