+++ src/YarpUtilities/include/BipedalLocomotionControllers/YarpUtilities/Helper.h
@@ -20,7 +20,6 @@
 
 #include <iDynTree/Core/VectorDynSize.h>
 
-
 namespace BipedalLocomotionControllers
 {
 
@@ -29,76 +28,6 @@ namespace BipedalLocomotionControllers
  */
 namespace YarpUtilities
 {
-
-/**
- * dependent_false is a type-dependent expression that is always false. Please check
- * https://en.cppreference.com/w/cpp/language/if for further details.
- */
-template <class T> struct dependent_false : std::false_type
-{
-};
-
-/**
- * is_iterable is used to build a type-dependent expression that check if an element is \a iterable
- * (i.e. the element has the methods <code>T::begin()<\code> and <code>T::end()<\code>). This
- * specific implementation is used when the the object is not iterable.
- */
-template <typename T, typename = void> struct is_iterable : std::false_type
-{
-};
-
-/**
- * is_iterable is used to build a type-dependent expression that check if an element is \a iterable
- * (i.e. the element has the methods <code>T::begin()<\code> and <code>T::end()<\code>). This
- * specific implementation is used when the the object is iterable, indeed
- * <code>std::void_t<\endcode> is used to detect ill-formed types in SFINAE context.
- */
-template <typename T>
-struct is_iterable<T, std::void_t<decltype(std::declval<T>().begin()), decltype(std::declval<T>().end())>>
-    : std::true_type
-{
-};
-
-/**
- * has_square_bracket_operator is used to build a type-dependent expression that check if an element
- * has <em>square bracket operator</em> (i.e. operator[]()). This specific implementation is used
- * when the the object does not have the square bracket operator
- */
-template <typename T, typename = void> struct has_square_bracket_operator : std::false_type
-{
-};
-
-/**
- * has_square_bracket_operator is used to build a type-dependent expression that check if an element
- * has <em>square bracket operator</em> (i.e. operator[]()). This specific implementation is used
- * when the the object has the square bracket operator, indeed <code>std::void_t<\endcode> is used
- * to detect ill-formed types in SFINAE context.
- */
-template <typename T>
-struct has_square_bracket_operator<T, std::void_t<decltype(std::declval<T>()[std::declval<int>()])>>
-    : std::true_type
-{
-};
-
-/**
- * is_resizable is used to build a type-dependent expression that check if an element is \a
- * resizable (i.e. the element has the methods <code>T::resize()<\code>). This specific
- * implementation is used when the the object is not \a resizable.
- */
-template <typename T, typename = void> struct is_resizable : std::false_type
-{
-};
-
-/**
- * is_resizable is used to build a type-dependent expression that check if an element is \a
- * resizable (i.e. the element has the methods <code>T::resize()<\code>). This specific
- * implementation is used when the the object is not \a resizable. Indeed
- * <code>std::void_t<\endcode> is used to detect ill-formed types in SFINAE context.
- */
-template <typename T>
-struct is_resizable<T, std::void_t<decltype(std::declval<T>().resize(std::declval<int>()))>> : std::true_type
-{
-};
 
 /**
  * Convert a value in a element of type T